Correction in reverting recommendations (#580)
* iam-recommender-at-scale first commit

* iam recommender at scale first commit

* adding install file

* updating the readme for installing iam-recommender-at-scale

* update tools with iam-recommender-at-scale

* update readme and install.sh

* update readme and install.sh

* update readme

* update errors

* correcting the naming and output messages

* changing the names from recommendation to recommendations

* correction in revert

* formatting modification

* removing unused json module
